Add support for malformed Windows file extensions
Windows file systems are case insensitive. Because of this in large archives you can often find some extensions that are in the wrong case and totally valid from an OS perspective. Just adding a tolower on the extensions before the check. Casting to a new var to keep the original as a constant.