Write REPLAYGAIN_REFERENCE_LOUDNESS
This provides programs with the necessary information to properly know
what the loudness of the files is. When not specified, -18 LUFS is often
assumed, so this is mainly important when a non-default --loudness
option is used. Nonetheless, writing the tag all the time (even for the
default -18 LUFS) allows players to be confident that they're achieving
a certain loudness, if that is desired.